feat (inbox): Add ability to mark posts complete / not complete
Each post now has a button allowing a user to mark the post as Done. If they click that, the post will only show up in the "Completed" filter. In that filter, the completed tasks can be marked uncompleted.

In this commit, the code to actually hit our server endpoints is commented out, and the completedByUser property is toggled purely on the frontend.